Try and integrate both ESP32 and ESP32S2 code into the one codebase...
Isn't going to work on Arduino. Platform IO only for the time being.

@vortigont - I've found a bug in this library introduced as a result of the fast functions. On the 'original' ESP32, remember how we have to 'flip' the order of uint16t's that go into the DMA buffer, because the TX FIFO flips them back again?

ESP32S2 doesn't have this stupid behaviour... but I noticed we flip the uint16t's correctly when we update RGB values, but we wern't going this when changing the control/address lines! I think this could have been a cause of ghosting and we didn't know why, because we were OE/LATing potentially earlier or later than required (or doing it multiple times).

Have a look at the MatrixPanel-I2S-DMA.cpp and see the crap I've introduced for the ESP32. It clearly wasn't a big issue as visually it was never noticed.